Handles deprecation. Fixes attributes
